BaSP-TG-32: setup week 10
