+++ jest.setup.js
@@ -1,0 +1,24 @@
+import mongoose from 'mongoose';
+import { MongoMemoryServer } from 'mongodb-memory-server';
+
+const mongoServer = new MongoMemoryServer();
+
+// Make the concection to a fake database
+
+beforeAll(async () => {
+  const mongo = await MongoMemoryServer.create();
+  const uri = mongo.getUri();
+
+  const mongooseOpts = {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  };
+  await mongoose.connect(uri, mongooseOpts);
+});
+
+// Disconnection the fake database
+afterAll(async () => {
+  await mongoose.connection.dropDatabase();
+  await mongoose.connection.close();
+  await mongoServer.stop();
+});
