+++ package.json
@@ -4,7 +4,7 @@
     "description": "app_name is a system that provides Human Resources services to companies.",
     "main": "index.js",
     "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
+        "test": "jest",
         "start": "nodemon --exec npm run babel-node -- src/index.js",
         "lint": "eslint .",
         "lint:fix": "eslint . --fix",
@@ -25,6 +25,7 @@
         "node": "14.x"
     },
     "dependencies": {
+        "dotenv": "^16.0.1",
         "express": "^4.18.0",
         "joi": "^17.6.0",
         "mongoose": "^6.3.3",
@@ -37,6 +38,9 @@
         "eslint": "^8.14.0",
         "eslint-config-airbnb-base": "^15.0.0",
         "eslint-plugin-import": "^2.26.0",
-        "husky": "^7.0.4"
+        "husky": "^7.0.4",
+        "jest": "^28.1.0",
+        "mongodb-memory-server": "^8.5.2",
+        "supertest": "^6.2.3"
     }
 }
