+++ src/index.js
@@ -1,33 +1,17 @@
 /* eslint-disable no-console */
-import express from 'express';
+import dotenv from 'dotenv';
 import mongoose from 'mongoose';
-import router from './routes';
+import app from './app';
 
 // Server
-const app = express();
+dotenv.config({ path: './.env' });
 const port = process.env.PORT || 3000;
-const mongoDBURL = 'mongodb+srv://BaSP_datatabase:RadiumRocket@basp-coco.zj49d.mongodb.net/BaSP-Coco?retryWrites=true&w=majority';
-
-// Body parser
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-// Routes
-app.use(router);
-
-app.get('/', async (req, res) => {
-  res.send('Hello World!');
-});
+const MONGO_URL = process.env.DATABASE_URL;
 
 // async function adminMod()
 
-app.listen(port, () => {
-  // eslint-disable-next-line no-console
-  console.log(`Example app listening on port ${port}`);
-});
-
 mongoose.connect(
-  mongoDBURL,
+  MONGO_URL,
   (error) => {
     if (error) {
       console.log('Fail to connect', error);
